Change API to require only advice, as a keyword argument #8 #9
This no longer allows passing a tuple of `(message, advice)` to an
assertion.  Instead, `advice` must be passed, as a keyword argument,
and the message is now optional and is named `msg` to match the
`unittest` API.

Additionally, this fixes #9 by allowing nested assertions to have
missing advice as long as the outer assertion has advice, which we
fall back to.